Added information on how to run tests in the README.md file.
This commit adds information on how to run the tests for the project in the README.md file. The unittest module built into Python is used to discover and run all test files in the tests directory that match the pattern *_test.py. The command to run the tests is provided, along with instructions on where to run the command from.