Add .gitignore
Add .gitignore file to project to exclude files and directories that should not be tracked by Git, such as compiled code, temporary files, and sensitive information. This helps keep the repository clean and reduces the risk of accidentally committing confidential information. The .gitignore file is based on the Python .gitignore template and includes common files and directories used in Python projects, as well as the project-specific files and directories that should be excluded.