Add DiGraphEx class and DAG constructor to dag_nx branch
The new DiGraphEx class extends the networkx DiGraph and provides additional methods to get root and leaf nodes and to perform a simple topological sort. The DAG class constructor now uses this extended graph class to build the graph and sequence order for the computation. In addition, a new method `assign_recursive_children_compound_priority` is added to calculate the compound priority of each node as the sum of the priorities of all children recursively. This method is called during DAG construction and sets the execution sequence order.

This commit only includes new code and does not modify any existing code.